adding the found substitutions to the formula given to the SAT solver
This is needed at least in two cases: 1) in LIA some equalities might not have a solution, e.g. 2x = 1; 2) In incremental mode it might happen that substitution is found in later frames for a variable present in previous frame. This means the inner solver already knows about the variable and will compute a value for it which might not be consistent with the substitution unless the solver is informed about this substitution