Prevent re-introduction of duplicate literals into clauses
Currently if a clause to be added to the SAT solver has duplicate literal which is know to be false, and we are in the interpolating mode, all occurences of such literal would be added back to the clause for proof logging purpose.

Since the literals are sorted, we can prevent this by checking if we have already encountered such literal before.