CoreSMTSolver: Generalize split clauses
For our version of model-based theory combination, TSolverHandler must
be able to communicate not just binary clause and not just single
clause, but multiple clauses that can be also ternary.
Concretly, for an interface equality x=y we want to create 3 clauses:
~(x = y) or x <= y
~(x = y) or y <= x
x = y or ~(x <= y) or ~(y <= x)

However, to maintain the SAT solver's invariants, we need to properly
handle backtracking and literal propagation from these clauses when they
are introduced.