Egraph: Introduce mechanism for dealing with dynamicly added terms
Egraph is sensitive to backtracking and needs to undo the made changes
exactly. The important thing is to maintain the invariants of signatures
and use-vectors. Adding new terms to Egraph after initialization causes
problems for maintaining the invariants when backtracking happens.

The solution is inspired by the approach of Yices. For these newly added
terms, we remember on the backtracking stack that we need to re-analyze
them on backtracking to maintain the invariants.

Re-analysis means that we remove these terms from the signature and
use-vectors data structures if they are the representative. After the
solver has backtracked, these terms are re-introduced to these
data-structures, if they are still the representative of their class.

One problem is that it is still not entirely clear to me at which point
we can stop re-analyzing these terms.