Interpret: Add model post-processing to avoid name clashes
Model validator does not like when a formal argument of function
definition has the same name as a variable already defined in the model.
To work around this limitation, this patch adds a small post-processing
that checks for such name clashes in the model and renames the formal
parameters in problematic function definitions.