LASolver: Prepare for adding new atoms on-the-fly
We aleady have a mechanism for adding new atoms that deals with LIA
splits. However, those atoms are special in the way that there are new
bounds on an existing variable.
This mechanism is generalized to allow bounds on arbitrary expressions
to be added dynamically.
The main issue is that new row being added to a tableaux must be
normalized, since he variables in the new expression might not be
nonbasic variables at the moment of addition.

Apart from that, we need to ensure that LASolver's data structure are
ready for dynamically added bounds of a new variable not seen at the
init stage.

The rest seems to be working nicely thanks for the mechanism of
QUASI_BASIC variables.