CoreSMTSolver: Log derivation of unit clauses from split clauses
Literals propagated at level 0 are special, the proof building code
assumes it has a unit clause as a reason. It is now possible that
literal can be propagated on level 0 from a split clause, so we need to
do the same trick as in regular SAT propagation.

This trick should be refactor to a separate method, probably...