CoreSMTSolver: Fix handling of split clauses
Previously if there were more than one propagation from split clauses on
level 0, we would remember only the last propagated literal and forget
the other ones. Moreover, the split clauses would not be added to the
SAT solver, thus losing information in the process.
This could lead to invalid models.