LASolver: Remove weird useless field
The field LASolver::cuts does not seem fulfil any useful purpose.
It should never happen that we would want to branch on the same constant
value multiple times. This would indicate a bug elsewhere in the code.
Moreover, the current handling of such situation, if it would occur, is
to silently ignore it, which is not a good approach.