CoreSMTSolver: Sort split clauses before processing
In order to avoid dealing with situation that a satisfied clause becomes
implying after backtracking because of some other implied clause, let's
sort the clauses first according to the last (highest) level where some literal
from the clause was assigned.
Then we can process the clauses starting from lowest last-assigned
level. This ensures that we need to backtrack at most once and avoids
the problem with satisfied clause becoming implied.