UFInterpolation: Fix memory leak
CGraph::edges is a vector of raw pointers, but it owns the objects.
The objects were correctly deleted when all edges are removed, but not
when a single edge is to be removed.

Later we can do a larger cleanup, but for now we simply delete the edge
before we remove it from our vector.