UFInterpolator: Fix interpolation in the present of distincts
UF theory conflict is caused by a disequality and a chain of (possibly
derived) equalities.
Previously, the code was assuming that it will find positive version of
the disequality in the labels map (unless the conflict is between two
interpreted constants). This assumption is not true if the disequality
is actually derived from a distinct term.
The fix is to consider also the possibility of presence of distinct
terms.