PartitionManager: Fix how partition of original formula is determined
Previously, we were relying on the partition information of the terms,
rather than on the index of the top-level formula.
This was not correct. Given a command `(assert b)`, `b` is a top-level
formula, say with index 1. Then if `(not b)` is present in another
top-level formula, say with index 2, then using term information, `b`
belongs to both partitons 1 and 2. But when we just need to partition
top-level formulas, `b` should have only index 1.

When determining A- and B-partitions for interpolation, we have to
consider only top-level partition indices, not partition information of
terms.