Interpol: Switch to internal interpolant validation
Switching to internal interpolant validation (instead of relying on
external decision procedure) allows us to enable interpolant validation
in our regression rest suite. Validating interpolants as part of the
regression test suite has great benefits.

OpenSMT's solver is mature enough so that we can trust it to verify
implications correctly. Typically, if there is problems with
interpolant, the bug is in the interpolation code, not in the solving
code.