Substitutor: Optimize substitution application to variables
The machinery of Rewriter is general, but also not optimal for simple
rewritings. Most importantly, it allocates a boolean mask (represented with
vector) to keep track of processed terms. This allocation is often
wasteful a should probably be re-thought.

Here we optimize Substitutor (a sub-class of Rewriter) to avoid the
general machinery when the term to process is a variable.
In this case, the general algorithm can be replaced by a simple lookup
in the substitutions map.