Logic+Rewriter: Add and use more persistent storage for marking terms
Instead of creating vector for marking terms on every call to a
rewriter, we can use permanent storage in Logic that will better re-use
the memory. It still need to grow when more terms are created, but it
should be able to re-use the storage without resize at least some times.
Moreover, using `nat_set` gives an ability to clear the auxiliary vector
in constant time, thus it is preferable over simple vector.