UFInterpolator: Add handling of a special case
When UF interpolation encounters an edge that is not colorable, it must
split the edge. So given an edge U -> V, it will find an AB-colorable term W that
is congruent with both U and V, and it will replace the edge with two
edges: U -> W -> V.
However, suppose that W had an edge W -> Z. Then in general, the new
path should be U -> W -> V -> Z.
But, there are two special cases: when Z = U and when Z = V.
The second special case has been handled in the code before. The first
one was not.
In this case, the last edge would complete a triangle loop U -> W -> V -> U.
Instead, the path should be just U -> W -> V, without the last edge
added. This commit does exactly that.