cmake: Error if static libraries not found
If the user requests a "maximally static build", give an error if the
libraries found do not match with `\.a$`.

In addition, disable from the build things that likely won't work: do
not build the dynamic library and do not build unit tests as they link
against the dynamic library.