Logic: Use wrapper with nicer interface for term marking
We can hide the implementation details of using nat_set and provide a
nicer interface (especially the constructor takes care of setting up the
low-level data structure) for term markings.

We can also mark the nat_set mutable (basically not considering it a
logical part of the Logic state) and thus allowing the method for
creating TermMarks helper constant on Logic.