LASolver: Avoid vectors with large gaps
When LAVarMapper is storing mapping from PTerms to LATerms as vectors
with TermId as index, this creates possibly huge vectors with just a
small number of entries filled. This is not a problem for single-use
instances, but in applications where number of terms in Logic is growing
constantly are a solver is queries repeatedly, the initialization of
LASolver due to allocation of large vectors starts to be noticeable.

For now we use standard hashmap as a replacement. It seems that some
kind of flat map, which additioanlly does not need to support deletion
could be even better.