Array: Basic lemma computation
We follow the implementation of SMTInterpol. This means we rebuild the
representation of the weakly-equivalence graph everytime there is a new
equality asserted.
When the representation is built, we collect the conditions for each
extended lemma.
Extended lemma says that if there are two selects a[i] and b[j] where i
is equivalent to j (i~j in Egraph) and a is weakly equivalent on i to b
then a[i] and b[j] must be equivalent.
Thus if we know that that there is a path from a to b that does not use
index i and all the stores use indices that *are known* to be different
from i, then the selects must be equal.
If the selects are known to be distinct, then there is a conflict.

The conditions of a lemma are thus the equality of the selects plus
equalities of all indices encountered on the path to index i.
If all these equalities are falsified, the lemma is a conflict lemma.

However, the full lemma must also explain not only the weakly equivalent
edges (caused by an array store), but also the strong equivalence edges
from Egraph.

What is still missing is the creation of new splits. When we are doing a
full check, we might not detect conflict, because an equality of indices
did not exist in the input formula, so the corresponding variable is
missing from the SAT solver.