TreeOps: a capped Pterm node counter
This commit introduces a method for counting the number of nodes in a
Pterm.  The use case is in estimating the printed sizes of Pterms when
serialising them, but the method could have other uses.

Given a term, the system computes number of subterms, or if the number
of subterms would be higher than a cap value m given at construction
time, a number that is at least m.

The system allows early exiting from the counting if the number of
subterms is greater than m, and supports caching.

The caching is based on the observation that if a subterm s of a term t
has at least n subterms, then t has at least n subterms.

In addition the system supports checking whether the cap value has been
reached.