Arrays: Keep track of asserted literals
Instead of relying on the information from Egraph, keep the track of
asserted literals directly in ArraySolver. The polarity map of Egraph
can contain additional information internally propagated in Egraph,
which the SAT solver does not yet now about.

Previously, the array solver could detect a conflict using information
from the Egraph, but the conflict clause was not really falsified yet in
the SAT solver and this was causing some problems. After this change,
the array solver basically waits until the information is propagated to
it thorugh the SAT solver.