CoreSMTSolver: Fix bug in handling split clauses
The invariant when adding new clauses to the SAT solver is that the
first literal must be unknown.
The split clause code did not maintain this in the case where there was
more than one unknown literal (and all others were falsified).

It is not clear to me if the second literal also needs to be unknown,
but it looks like that is not necessary.