Splitter: do not expose internal vars
With this commit clause sharing and splitting run the reverse aux var
rewriter to PTRefs that are exposed to the outside.  As a result no
variables prefixed with `.ite`, `.div`, or `.mod` should appear in the
shared clauses or splits.

Note: The dag counter is issued twice for the PTRefs to be shared: once
to see if the original PTRef is too deep to be traversed for removal of
aux vars, and second time to see if the aux-expanded PTRefs are too big
to be shared.

Co-authored-by: Masoud Asadzadeh <masoud.asadzade@gmail.com>