Egraph: Remove an assert
As a result of the remuval of `isKnown` and changing the uses to
`isInformed`, this assertion no longer holds.

There is no change in what Enodes are created. This happens for
equalities inside an uninterpreted function.  The change is that before
such equalities were marked as known but not as informed.  The code is
`Egraph::getPolaritySuggestion` checks informed. Thus previously the
execution stopped at this check, but now it gets past it.

Co-authored-by: Martin Blicha <martin.blicha@gmail.com>