TSolvers: remove isKnown
Currently there are two ways of determining whether a term is known to a
theory solver: `isInformed` and `isKnown`.  These maintain more or less
the same information, and therefore the system constitutes a potential
source of inconsitency and ultimately bugs.  This commit simplifies the
code by removing `isKnown` and substituting its uses by `isInformed`.

The relevant interface now has two methods: `isValid` which returns `true`
if the term can be declared to the theory solver, and `isInformed` which
returns `true` if the term has been declared to the theory solver.

As a side effect this commit fixes a bug that causes the difference
logic solvers to crash in incremental solving.