ArithLogic: Rewrite substitutions
This introduces two special phases before falling back to the previous
behaviour.
First, it tries to apply as much substitutions by a constant as possible,
iterating until fixed point. This is cheaper that doing this in multiple
iterations of the main substitution loop.

Second, it searches for variables present only in a single equality. This
equality is then used to derive a substitution for such variable.