TSolverHandler: follow solver schedule
Currently the order dictated in which solvers should be accessed, e.g.,
when asserting literals, is defined in a vector
`TSolverHandler::SolverSchedule`.  However, the code does not respect
this order.  Instead the code uses an order that is non-deterministic
between systems as it depends on in which order certain static data
structures are initialised.

This commit fixes the code to follow strictly the order from
`TSolverHandler::SolverSchedule`.