LASolver: Set `has_explanation` flag properly
It seems that LASolver was always setting `has_explanation` flag to
true, but never to false. This, in combination with inverted solver
schedule in UFLA, led to a subtle bug where UF would detect conflict,
but LA solver would also signal that it has a conflict (even though this
would not be true) and then return an empty explanation. This resulted
in incorrect UNSAT answers.