ArithLogic: Fix substitutions in presence of UF
Previously, in `collectSingleEqualitySubstitutions`, the variables were
processed in order as stored in `unordered_map`. Additionally, UF terms
are treated as variables in ArithLogic. This meant that from an equality
f(x) = x (with x being Real), the code could derive a substitution "x ->
f(x)". However, such substitution results in an infinite cycle in the
function that computes transitive closure of substitutions.

We fix this by forcing an order in which variables are checked, starting
from variables with higher PTRef, which mean they they were created
later. This ensures for the above case, that f(x) will be considered
before x, so the computed substitution will be "f(x) -> x".