Everywhere: Fix weirdness in PTRef comparison
Previously, PTRef was exposing operator< which was, however, using
operator> on the underlying uint32_t. This is extremely
counterintuitive.
This commit makes a first step towards normalizing comparison of PTRefs.
To preserve the current behaviour, it changes the operator< to operator>
and modifies the code throughout the codebase that previously
dependended on operator< to now use operator>. As the implementation
stays the same, this should not change any observable behaviour.