ArithLogic: Simplify method SimplifyConst::simplifyConstOp
The first branch before the change did not make sense, because it was
trying to get sort from the first term in an empty list.
Since this method is internal to ArithLogic.cc, we have control about
how it is called. It is better to just assert that the input will
never be an empty list.