ArithLogic: Avoid VecMap usage in mkPlus
In application that issues a large number of simple queries, such as
Spacer algorithm, a surprisingly large portion of the runtime is spend
in ArithLogic::mkPlus.

One reason is the usage of VecMap which is quite memory-intensive data
structure.  With the proposed change we hope to reduce the amount of
memory allocated and deallocated in each call to mkPlus.  Less work with
memory should lead to better overall performance.