interpret: Improve let term parsing
Current master checks for redundant "parallel" let bindings of the form
`(let ((l a) (l b)) f)` and gives an error if one is present.

The smt-lib standard does not forbid these constructs and other solvers
seem to parse them without errors.

Hence one can argue that the checking in opensmt unnecessarily
introduces complication in the code and in principle also a performance
overhead (I haven't measured this)

This commit introduces a semantic in which the last binding prevails
(in the example above `(l b)`).