ArithLogic: Avoid redundant work in computing substitutions
Equalities are currently normalized to form c = t where c is a constant
and t is a term.
When computing polynomial from the equality, we compute the term t - c.
However, mkMinus works currently by negating the SECOND argument and
passing them to mkPlus.
It is better to negate the constant rather than the term, thus we should
pass FIRST argument of the equality as the SECOND argument to mkMinus.