UFLA: Create interface lemmas if required by array solver
It may happen that the array solver, as part of its lemmas, requires the
creation of an equality between two interface variables. In this case,
we need to immediately create corresponding interface lemmas, otherwise
such lemmas will not be created and the solver may incorrectly assert
inequality in arithmetic solver and equality in egraph.

The current solution is quite hacky, it would be better to find a
different solution to this problem.