Arrays: Fix missing literals in read-over-weak-eq lemmas
When computing the explanation for an array conflict (read-over-weak-eq
lemma), we also need to add the conditions that explain why the select
index is equivalent to its root.
Thus might not be always true, so for the lemma to be valid, this
explanation must be included.

Looks like we are already doing this for the extensionality lemmas, so
we probably hit this problem before there.