Interpret: Reduce size of call stack's activation record for parseTerm
Last changes in parseTerm (a838804c) led to increased size of activation
record and this resulted in more stack overflows on SMT-LIB benchmarks
with huge nested lets.
Even though according to my intuition, the abovementioned changes should
decrease the memory footprint on the stack, it seems the opposite
happened. My guess is that the changes allowed compiler to inline
something that was not inlined previously, but this required more stack
space to accomodate the inlined code.

The changes in this commit move more work (and data structure) to
another method, which is then not inlined anymore (my guess). This not
only reverts the size increase, but actually leads to further decrease
of the activation record's size.

Proper solution would be to reimplement parsing in a non-recursive way,
but this patch significantly helps in the meantime.