LASolver: An attempt to find better cuts from proofs
On the simple example from SMT-LIB (QF_LIA/check/int_incompleteness3.smt2),
I observed that cuts from proofs always were always returning a cut
around a single variable, effectively doing the same as the simple
branching would do. I also noticed that the behaviour on this example
changed between v2.4.2 and v2.4.3. In v2.4.2, this example was solved
immediately, because cuts from proof actually came up with an
interesting explanation involving two variables, not just a single one.
Further analysis showed that the Hermite normal form that we compute,
and hence the proof of unsatisfiability discovered depends on the
ordering of the constraints. This dependence on the order of the
constraints is understable, but on this example it had the unfortunate
effect that it basically degraded to the behaviour of simple
branch-and-bound.

I am not sure if there is something wrong with our implementation, but
as a quick workaround, I suggest to sometimes change the order of
constraints.
I was considering std::shuffle, but that requires random generator which
is implementation-defined. To be consistent across compilers and
standard libraries, I suggest instead to sometimes reverse the
constraints, using our own random test.

This solves the divergence on this particular example.