ArithLogic: Fix computation of substitutions
Since we use transitive closure when computing substitutions, we should
never start with a substitution "x -> t(x)" where we want to substitute
variable x with a term containing this variable. Such a substitution
leads to an inifinite rewriting in our code in transitive closure
computation.

Since the terms in our polynomials are ordered by increasing PTRef of
term's variable, we need to take the largest one, not smallest one when
choosing which variable to use for the substitution.
Otherwise we might end up in the situation described above.