ArithLogic: Limit substitutions to prevent term explosion
We have observed problems on some subcategories of QF_LIA benchmarks
(20220307-SMPT, mainly Shared-Memory*) with our rewriting based on
substitutions. The problem is that those benchmarks use variables to
encode subsums of sums of large number of variables. With our previous
approach, we would try to replace most of these variables with the
expanded sum they represent. The often led to blowup in the number and
size of terms we were creating.

Here we apply a limit for the substitutions we apply, namely we only
allow arithmetic substitutions that do not increase the number of
variables. In other words we only allow to replace a variable with a
constant or other variable, possibly with additional constant term; the
substituted value must be of a form ax + c, with a and c possibly 0.