Interpret: Fix treatment of define funs
Previously, we created variables corresponding to the formal arguments
of defined functions (define-fun command). However, since in OpenSMT, we
store variables with its name and sort, this led to clashes when a
proper variable with the same name but different sort was declared
later. However, the names of formal arguments of define-fun should not
prevent a declaration of a variable with the same name.

The solution proposed here is to use unique internal names to store the
formal arguments, while using our existing let mechanism to introduce
the actual names used in the script to refer to these internal
variables.