Interpet: Fix get-value response
The response to get-value command is a sequence of pairs where the first
element of each pair should be the term as given in the command (the
second element is its value in the model).
Therefore, we should not try to resolve the given term, as is the
current behaviour. Instead, we can just print the string parsed back to
the user.