Fix assertion failure in `QF_UFLIA` with `mod`/`div`
Currently `QF_UFLIA` theory not call `DivModRewriter` on the formula.
As a result, the arithmetic solver sees, e.g., `mod` terms, and crashes
on an assertion failure.  In optimized code the error results in the
worst case in unsoundness.

This commit adds the necessary rewrites to `UFLATheory.cc`, as well as a
missing call to rewriter for `distinct`s.  The latter is untested.