Change architecture around assertion stack, preprocessing and cnfization
Maintaining the assertion stack should be the responsibility of the
MainSolver, not the Theory.
Related to this change of responsibilities, we also change the
preprocessing, so that it operates only on the current frame.
(This should probably be changed so that simplifications from lower
frames can be applied to higher frames as well, but we need to test if
this would really be useful.)

Another change is a more separate Cnfizer, which now does not need to
know about the PartitioningManager or the CoreSolver. Instead, it just
reports new clauses to a callback that needs to be set up beforehand.

The larger goal here is to eventually remove separate Theory and Logic classes and
have a single class responsible for managing terms (former Logic) and
a single class for preprocessing (former Theory). These could be
customized using a plugin architecture.