Logic: Make test for quotes more transparent
Instead of using a hard-to-read table, let's follow the standard to the
letter. The standard defines which characters are permitted in the
simple symbol (which does not require quotes). If we encounter a
character not in the specified set, we need to add quotes.

The reason for this change has been moivated by compilation problem in
JavaSMT and the fact that even though the table was created with length
256, only the first 250 entries were specified.