Rewriting: Do not collect rewritten arithmetic inequalities
The information about which inequalities originated from arithmetic
equalities were recorded during preprocessing. However, this makes it
difficult to separate preprocessing passes and make them more
independent, i.e., so we can run them on single formula as well as on
group of formulas.
If this information is needed, it should be obtain in a different way.
At the moment, this information is only used by the lookahead solver for
partitioning, where the strategy is to not split on such inequalities.
However, I do not see benefit in such strategy.
Since the need to record this information is blocking further
refactoring that has the potential to unlock performance improvements in
incremental solving, I propose to just not record this information
anymore.