ArithLogic: Fix handling of minus operator with more than two arguments
According to the SMT-LIB definition of theory of Ints, operator minus is
left associative. This means that it is allowed to create a function
application of minus to more than two arguments. This is than a
syntactic sugar that applies the operator recursively first to all the
arguments except the last one and then apply the operator to the result
and the last argument.