ProofGraph: Simplify computation of topological order
This PR replaces the current implementation of top-down (parents before
child) topological order for resolution proof. New implementation uses a
simple DFS traversal from root, marking child only after its parents are
processed.
The advantage of the new implementation is that it does not need the
information about resolvents of each proof node.
Regarding performance, anectodal evidence from Golem shows that the new
implementation performs better, because it does not need to traverse
resolvents of each node, which is a std::set.