ITEs: Avoid allocating vector with size equal to number of all terms
While in a single-query-solving, allocating a vector with size equal to
the number of all known terms is not a problem, in a setting where we
are solving many small queries and creating huge number of terms, this
was observed to actually impact performance.

Specifically, the backtranslation of proof in Golem invokes OpenSMT many
times when it is recomputing missing steps (when predicates are
eliminated during preprocessing).
In one case, this change eliminated approximately 10 seconds of proof
production overhead to almost zero.