Core: Add missing derivation of unit clauses at level 0 for splits
When a clause becomes unit under current assignment at level 0, we need
to create the actual unit clause and log its derivation in the
resolution proof, otherwise, the final proof would not be constructed
correctly.

We have already been doing that in CoreSMTSolver::propagate, but it was
missing in CoreSMTSolver::handleNewSplitClauses. We refactor this
functionality to a separate method and call it also in the second case.