Arrays: Fix unsoundness in handling nested arrays
As part of the initialization of the array solver, we need to collect
all array, store and select terms.
Previously, we incorrectly assume that array terms and select terms are
mutually exclusive. This is not true in case of nested arrays when a
select term can be of an array sort.
For this reason we previously missed some select terms and consequently
did not add corresponding lemmas to the solver.